// src/models/IBreed.ts
export interface IBreed {
  name: string;
  id: string;
  cardDesc: string;
  cardImg: string;
  headImg: string;
  subDesc: string;
  desc: string;
  history: string;
  temperament: string;
  health: string[];
  size: number;
  energyLvl: number;
  exReqs: number;
  playfulness: number;
  affLvl: number;
  frToDogs: number;
  frToPets: number;
  frToStr: number;
  watchfulness: number;
  easeTr: number;
  groomReqs: number;
  heatSens: number;
  vocality: number;
}